requirements.txt
# Flask
Flask==0.11
flask-restful==0.3.0
Flask-HTTPAuth==2.2.0
flask-bcrypt==0.6.0

# Flask Extensions
WTForms>=2.1
Flask-Assets==0.10
Flask-Cache==0.13.1
Flask-DebugToolbar==0.9.0
Flask-Login==0.3.2
Flask-Script==2.0.5
Flask-WTF==0.10.0
Flask-PyMongo >= 0.4.1
Werkzeug==0.11.10
Flask-Bootstrap==3.1.1.1
Jinja2==2.8
Flask-Moment==0.3.0

jsonpickle
pyjwt

# Database
pymongo>=3.2.2